InputValue.FLOAT -> InputValue.DECIMAL

+++ common/src/main/java/mcp/mobius/waila/config/ConfigEntry.java
@@ -13,7 +13,7 @@ public class ConfigEntry<T> {
 
     public static final ConfigValueFactory<Boolean> BOOLEAN = BooleanValue::new;
     public static final ConfigValueFactory<Integer> INTEGER = (n, d, s) -> new InputValue<>(n, d, s, InputValue.INTEGER);
-    public static final ConfigValueFactory<Double> DOUBLE = (n, d, s) -> new InputValue<>(n, d, s, InputValue.FLOAT);
+    public static final ConfigValueFactory<Double> DOUBLE = (n, d, s) -> new InputValue<>(n, d, s, InputValue.DECIMAL);
     public static final ConfigValueFactory<String> STRING = InputValue::new;
     public static final ConfigValueFactory<Enum<?>> ENUM =  (n, d, s) -> new EnumValue(n, d.getDeclaringClass().getEnumConstants(), d, s);
